FIRST PART OF THE LANDING PAGE
I have made the first part of the website according to the specifications given.
The readme file has been updated to contain credits for images used and main
objective of the project.
An images folder which will contain all images to be used in the page, an index
html page containing the pages html and a style css page containing the pages
css code have also been added to the project.